Adding code to make a interface with Jython
Using Jython we can not call Python C/C++ API which we used to implement
the search on the graph. So we have to work around on it, it nows see if
the interpreter is CPython or Jython, if it is CPython the result of
gsflc.search is writer into a file with the same name of the lattice, but
the extension that is '.set', and then when the code is executed with
Jython it just read this '.set', instead of call gsflc.search.

That is, to use Jython, we first need to run with CPython (something
that call search_sentences_c).