Adding codo to call weka and make re_score.
With GsflReScore we can re score the sentence using a set of hypothesis
(detecting confusion only between two phonemes). When it found a
possible confusion, it calls weka to classify it as confusion or not, if
classified as confusion a gain is given to a determined phoneme
otherwise it is just ok.

To run it you need weka.jar in your class path and Jython interpreter.